api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
kafka-python==2.0.2
